Update allowing for action based responses
This update allows questions to be answered through pre-programmed processes instead of only being able to respond to questions asking for facts. A new file called jarvis_actions was created to store the functions that perform specific actions to respond to questions and/or find an answer to the user's question.