Added the capability to return several answers.
This update allows the user to ask questions that require more than one answer to properly respond. This updates how the jarvis.py file calls the Jarvis_Memory class. This also makes the Jarvis_Memory class split the data_points object into phrase_data_points which stores the data_points lists for each phrase in the input. This allows questions like: "what is your name and my name?" to be asked and properly handled by returning two names in the order the questions were asked. i.e. In response to the example above, if my name is Jimmy then it returns ['Jarvis', 'Jimmy']